habit_tracker/requirements.txt
matplotlib>=3.7.1,<4.0.0
ttkthemes>=3.2.2
pillow>=9.5.0
